removed warnings
git-svn-id: http://svn.seqan.de/seqan/trunk@9170 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/variant_comp/variant_comp.cpp
@@ -514,9 +514,9 @@ int main(int argc, const char *argv[])
 					while (!_streamEOF(file))
 					{
 						parseSkipWhitespace(file,c);
-						int rangeBegin = parseReadDouble(file,c);
+						int rangeBegin = static_cast<int>(parseReadDouble(file,c));
 						parseSkipWhitespace(file,c);
-						int rangeEnd = parseReadDouble(file,c);
+						int rangeEnd = static_cast<int>(parseReadDouble(file,c));
 						appendValue(options.ranges,Pair<int,int>(rangeBegin,rangeEnd));
 					}
 					file.close();
